refactor: Refactoring for Object-Oriented Design
