+++ src/main/java/com/mcc/outify/weathers/WeatherAPI.java
@@ -1,0 +1,9 @@
+package com.mcc.outify.weathers;
+
+import java.io.IOException;
+
+public interface WeatherAPI {
+
+    void getWeatherData(String latitude, String longitude) throws IOException;
+
+}
