+++ src/main/java/com/mcc/outify/weathers/weatherApiImpl/YrAPI.java
@@ -1,28 +1,27 @@
-package com.mcc.outify.apis;
+package com.mcc.outify.weathers.weatherApiImpl;
 
+import com.mcc.outify.weathers.WeatherAPI;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.io.BufferedReader;
+import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLEncoder;
 
 @Component
-public class Yr {
+public class YrAPI implements WeatherAPI {
 
     @Value("${yr-user-agent}")
     private String yrUserAgent;
 
-    public void run() throws Exception {
-
-        String lat = "60.10";
-        String lon = "9.58";
-
+    @Override
+    public void getWeatherData(String latitude, String longitude) throws IOException {
         StringBuilder urlBuilder = new StringBuilder("https://api.met.no/weatherapi/locationforecast/2.0/complete");
-        urlBuilder.append("?" + URLEncoder.encode("lat", "UTF-8") + "=" + URLEncoder.encode(lat, "UTF-8"));
-        urlBuilder.append("&" + URLEncoder.encode("lon", "UTF-8") + "=" + URLEncoder.encode(lon, "UTF-8"));
+        urlBuilder.append("?" + URLEncoder.encode("lat", "UTF-8") + "=" + URLEncoder.encode(latitude, "UTF-8"));
+        urlBuilder.append("&" + URLEncoder.encode("lon", "UTF-8") + "=" + URLEncoder.encode(longitude, "UTF-8"));
         URL url = new URL(urlBuilder.toString());
         HttpURLConnection conn = (HttpURLConnection) url.openConnection();
         conn.setRequestMethod("GET");
@@ -41,6 +40,8 @@ public class Yr {
         }
         rd.close();
         conn.disconnect();
+
+        // 아래 print문 삭제 후, DB 저장 로직 구현 필요
         System.out.println(sb);
     }
 }
